add wizard header exit util
